cleaning up PR
